consensus: KORE Protocol V3 Full 256-bit Stake Modifier Upgrade
Implement complete 256-bit stake modifier algorithm to eliminate stake
grinding attacks. This upgrade expands the modifier search space from
2^64 to 2^256, making pre-computation attacks computationally infeasible.

Changes:
- Add ComputeStakeModifierV2() returning full uint256 instead of truncated
  uint64_t for V3+ blocks
- Add ConvertModifier64To256() helper for fork boundary transition
- Add nStakeModifierV2 field to CBlockIndex for 256-bit storage
- Update GetKernelStakeModifier() to return uint256
- Update CheckStake() to use 256-bit modifier in hash computation
- Update Stake() and CheckProofOfStake() for 256-bit support
- Configure V3 fork heights: Mainnet 2439950, Testnet 200, Regtest 0
- Add serialization support for nStakeModifierV2 in CDiskBlockIndex
- Remove static PREDEFINED_MODIFIER fallback for V3 blocks (returns error)

Fork Boundary Behavior:
- At V3 fork height, previous 64-bit modifier is converted to 256-bit
- All subsequent V3+ blocks use full 256-bit modifier chain
- Legacy 64-bit field populated with low 64 bits for backward compatibility

Security: Eliminates stake grinding vulnerability by requiring 2^256 work
to manipulate modifier outputs. Pre-computation attacks become impossible.

Note: This commit implements Option B from the V3 continuation plan,
extending 256-bit modifiers to ALL V3+ blocks (not just fork boundary).

+++ src/stakeinput.cpp
@@ -113,7 +113,8 @@ bool CKoreStake::CreateTxOut(CWallet* pwallet, CTxOut& txOut)
     return true;
 }
 
-bool CKoreStake::GetModifier(uint64_t& nStakeModifier)
+// V3: Returns 256-bit stake modifier
+bool CKoreStake::GetModifier(uint256& nStakeModifier)
 {
     GetIndexFrom();
     if (!pindexFrom)
