+++ src/stakeinput.h
@@ -23,7 +23,7 @@ public:
     virtual CAmount GetValue() = 0;
     virtual bool CreateLockingTxOuts(CWallet* pwallet, vector<CTxOut>& vout, CAmount value) = 0;
     virtual bool CreateTxOut(CWallet* pwallet, CTxOut& txOut) = 0;
-    virtual bool GetModifier(uint64_t& nStakeModifier) = 0;
+    virtual bool GetModifier(uint256& nStakeModifier) = 0;  // V3: 256-bit modifier
     virtual uint256 GetOldModifier(bool isProofOfStake) = 0;
     virtual CDataStream GetUniqueness() = 0;
     virtual int GetPosition() = 0;
@@ -49,7 +49,7 @@ public:
     CBlockIndex* GetIndexFrom() override;
     bool GetTxFrom(CTransaction& tx) override;
     CAmount GetValue() override;
-    bool GetModifier(uint64_t& nStakeModifier) override;
+    bool GetModifier(uint256& nStakeModifier) override;  // V3: 256-bit modifier
     uint256 GetOldModifier(bool isProofOfStake) override;
     CDataStream GetUniqueness() override;
     bool CreateTxIn(CWallet* pwallet, CTxIn& txIn, uint256 hashTxOut = 0) override;
